refactored unit-tests and updated spring
