--- backend/docs/elysianmuse/org.darkSolace.muse.userModule.model/-suspension-history/-suspension-history.md
@@ -1,11 +1,0 @@
-//[elysianmuse](../../../index.md)/[org.darkSolace.muse.userModule.model](../index.md)/[SuspensionHistory](index.md)
-/[SuspensionHistory](-suspension-history.md)
-
-# SuspensionHistory
-
-[jvm]\
-fun [SuspensionHistory](-suspension-history.md)(
-id: [Long](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-long/index.html)?, user: [User](../-user/index.md),
-suspendedDate: [Date](https://docs.oracle.com/javase/8/docs/api/java/util/Date.html) = Date(),
-confirmationCode: [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-string/index.html) = UUID.randomUUID()
-.toString(), acceptedDate: [Date](https://docs.oracle.com/javase/8/docs/api/java/util/Date.html)? = null)
