--- backend/docs/elysianmuse/org.darkSolace.muse.userModule.model/-suspension-history/accepted-date.md
@@ -1,7 +1,0 @@
-//[elysianmuse](../../../index.md)/[org.darkSolace.muse.userModule.model](../index.md)/[SuspensionHistory](index.md)
-/[acceptedDate](accepted-date.md)
-
-# acceptedDate
-
-[jvm]\
-var [acceptedDate](accepted-date.md): [Date](https://docs.oracle.com/javase/8/docs/api/java/util/Date.html)? = null
