--- backend/docs/elysianmuse/org.darkSolace.muse.userModule.model/-suspension-history/confirmation-code.md
@@ -1,7 +1,0 @@
-//[elysianmuse](../../../index.md)/[org.darkSolace.muse.userModule.model](../index.md)/[SuspensionHistory](index.md)
-/[confirmationCode](confirmation-code.md)
-
-# confirmationCode
-
-[jvm]\
-val [confirmationCode](confirmation-code.md): [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-string/index.html)
