--- backend/docs/elysianmuse/org.darkSolace.muse.userModule.model/-suspension-history/id.md
@@ -1,7 +1,0 @@
-//[elysianmuse](../../../index.md)/[org.darkSolace.muse.userModule.model](../index.md)/[SuspensionHistory](index.md)
-/[id](id.md)
-
-# id
-
-[jvm]\
-val [id](id.md): [Long](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-long/index.html)?
