--- backend/docs/elysianmuse/org.darkSolace.muse.userModule.model/-suspension-history/index.md
@@ -1,21 +1,0 @@
-//[elysianmuse](../../../index.md)/[org.darkSolace.muse.userModule.model](../index.md)/[SuspensionHistory](index.md)
-
-# SuspensionHistory
-
-[jvm]\
-@Entity
-
-class [SuspensionHistory](index.md)(id: [Long](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-long/index.html)?,
-user: [User](../-user/index.md), suspendedDate: [Date](https://docs.oracle.com/javase/8/docs/api/java/util/Date.html),
-confirmationCode: [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-string/index.html),
-acceptedDate: [Date](https://docs.oracle.com/javase/8/docs/api/java/util/Date.html)?)
-
-## Properties
-
-| Name | Summary |
-|---|---|
-| [acceptedDate](accepted-date.md) | [jvm]<br>var [acceptedDate](accepted-date.md): [Date](https://docs.oracle.com/javase/8/docs/api/java/util/Date.html)? = null |
-| [confirmationCode](confirmation-code.md) | [jvm]<br>val [confirmationCode](confirmation-code.md): [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-string/index.html) |
-| [id](id.md) | [jvm]<br>val [id](id.md): [Long](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-long/index.html)? |
-| [suspendedDate](suspended-date.md) | [jvm]<br>val [suspendedDate](suspended-date.md): [Date](https://docs.oracle.com/javase/8/docs/api/java/util/Date.html) |
-| [user](user.md) | [jvm]<br>val [user](user.md): [User](../-user/index.md) |
