--- backend/docs/elysianmuse/org.darkSolace.muse.userModule.model/-suspension-history/suspended-date.md
@@ -1,7 +1,0 @@
-//[elysianmuse](../../../index.md)/[org.darkSolace.muse.userModule.model](../index.md)/[SuspensionHistory](index.md)
-/[suspendedDate](suspended-date.md)
-
-# suspendedDate
-
-[jvm]\
-val [suspendedDate](suspended-date.md): [Date](https://docs.oracle.com/javase/8/docs/api/java/util/Date.html)
