--- backend/docs/elysianmuse/org.darkSolace.muse.userModule.model/-suspension-history/user.md
@@ -1,7 +1,0 @@
-//[elysianmuse](../../../index.md)/[org.darkSolace.muse.userModule.model](../index.md)/[SuspensionHistory](index.md)
-/[user](user.md)
-
-# user
-
-[jvm]\
-val [user](user.md): [User](../-user/index.md)
