--- backend/docs/elysianmuse/org.darkSolace.muse.userModule.repository/-suspension-history-repository/get-by-id.md
@@ -1,9 +1,0 @@
-//[elysianmuse](../../../index.md)/[org.darkSolace.muse.userModule.repository](../index.md)
-/[SuspensionHistoryRepository](index.md)/[getById](get-by-id.md)
-
-# getById
-
-[jvm]\
-abstract fun [getById](get-by-id.md)(
-id: [Long](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-long/index.html)): [SuspensionHistoryEntry](../../org.darkSolace.muse.userModule.model/-suspension-history-entry/index.md)
-?
