+++ backend/docs/elysianmuse/org.darkSolace.muse.userModule.service/-user-role-service/index.md
@@ -33,7 +33,7 @@ jvm
 | [changeRole](change-role.md) | [jvm]<br>@Transactional<br>fun [changeRole](change-role.md)(user: [User](../../org.darkSolace.muse.userModule.model/-user/index.md), role: [Role](../../org.darkSolace.muse.userModule.model/-role/index.md)): [User](../../org.darkSolace.muse.userModule.model/-user/index.md)?<br>Changes the role of a [User](../../org.darkSolace.muse.userModule.model/-user/index.md) and persists it in the database. |
 | [confirmSuspension](confirm-suspension.md) | [jvm]<br>fun [confirmSuspension](confirm-suspension.md)(confirmationCode: [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-string/index.html)): [Boolean](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-boolean/index.html)<br>Confirms the acceptance of a [SuspensionHistoryEntry](../../org.darkSolace.muse.userModule.model/-suspension-history-entry/index.md), identified by its confirmation code. |
 | [getAllCurrentlySuspendedUsers](get-all-currently-suspended-users.md) | [jvm]<br>fun [getAllCurrentlySuspendedUsers](get-all-currently-suspended-users.md)(): [List](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin.collections/-list/index.html)&lt;[User](../../org.darkSolace.muse.userModule.model/-user/index.md)&gt;<br>Retrieves all currently suspended users |
-| [getSuspensionHistory](get-suspension-history.md) | [jvm]<br>fun [getSuspensionHistory](get-suspension-history.md)(id: [Long](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-long/index.html)): [List](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin.collections/-list/index.html)&lt;[SuspensionHistoryEntry](../../org.darkSolace.muse.userModule.model/-suspension-history-entry/index.md)&gt;<br>Retrieves the suspension history of a given user, identified by its id.<br>[jvm]<br>fun [getSuspensionHistory](get-suspension-history.md)(user: [User](../../org.darkSolace.muse.userModule.model/-user/index.md)): [List](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin.collections/-list/index.html)&lt;[SuspensionHistoryEntry](../../org.darkSolace.muse.userModule.model/-suspension-history-entry/index.md)&gt;<br>Retrieves the suspention history of a given user. |
+| [getSuspensionHistory](get-suspension-history.md) | [jvm]<br>fun [getSuspensionHistory](get-suspension-history.md)(id: [Long](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-long/index.html)): [List](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin.collections/-list/index.html)&lt;[SuspensionHistoryEntry](../../org.darkSolace.muse.userModule.model/-suspension-history-entry/index.md)&gt;<br>Retrieves the suspension history of a given user, identified by its id.<br>[jvm]<br>fun [getSuspensionHistory](get-suspension-history.md)(user: [User](../../org.darkSolace.muse.userModule.model/-user/index.md)): [List](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin.collections/-list/index.html)&lt;[SuspensionHistoryEntry](../../org.darkSolace.muse.userModule.model/-suspension-history-entry/index.md)&gt;<br>Retrieves the suspension history of a given user. |
 | [suspendUser](suspend-user.md) | [jvm]<br>@Transactional<br>fun [suspendUser](suspend-user.md)(id: [Long](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-long/index.html)): [User](../../org.darkSolace.muse.userModule.model/-user/index.md)?<br>Suspends the provided [User](../../org.darkSolace.muse.userModule.model/-user/index.md), identified by its id, and persists it in the database. Also creates a [SuspensionHistoryEntry](../../org.darkSolace.muse.userModule.model/-suspension-history-entry/index.md) if it doesn't exist.<br>[jvm]<br>@Transactional<br>fun [suspendUser](suspend-user.md)(user: [User](../../org.darkSolace.muse.userModule.model/-user/index.md)): [User](../../org.darkSolace.muse.userModule.model/-user/index.md)?<br>Suspends the provided [User](../../org.darkSolace.muse.userModule.model/-user/index.md) and persists it in the database. |
 | [suspensionCodeForUsername](suspension-code-for-username.md) | [jvm]<br>fun [suspensionCodeForUsername](suspension-code-for-username.md)(username: [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-string/index.html)): [String](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-string/index.html)?<br>Retrieves the confirmation code for a users open [SuspensionHistoryEntry](../../org.darkSolace.muse.userModule.model/-suspension-history-entry/index.md). The user is identified by its username. |
