+++ backend/src/test/kotlin/org/darkSolace/muse/securityModule/AuthenticationServiceTests.kt
@@ -1,26 +1,19 @@
 package org.darkSolace.muse.securityModule
 
-import org.darkSolace.muse.DBClearer
 import org.darkSolace.muse.securityModule.model.LoginRequest
 import org.darkSolace.muse.securityModule.model.SignUpRequest
 import org.darkSolace.muse.securityModule.model.SignUpResponse
 import org.darkSolace.muse.securityModule.service.AuthenticationService
+import org.darkSolace.muse.testUtil.TestBase
 import org.darkSolace.muse.userModule.model.User
 import org.darkSolace.muse.userModule.service.UserRoleService
 import org.darkSolace.muse.userModule.service.UserService
-import org.junit.jupiter.api.*
+import org.junit.jupiter.api.Assertions
+import org.junit.jupiter.api.Order
+import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.test.context.SpringBootTest
-import org.springframework.test.context.DynamicPropertyRegistry
-import org.springframework.test.context.DynamicPropertySource
-import org.testcontainers.containers.PostgreSQLContainer
-import org.testcontainers.junit.jupiter.Container
-import org.testcontainers.junit.jupiter.Testcontainers
 
-@SpringBootTest
-@Testcontainers
-@TestMethodOrder(MethodOrderer.OrderAnnotation::class)
-class AuthenticationServiceTests {
+class AuthenticationServiceTests : TestBase() {
     @Autowired
     lateinit var authService: AuthenticationService
 
@@ -29,32 +22,6 @@ class AuthenticationServiceTests {
 
     @Autowired
     lateinit var userRoleService: UserRoleService
-
-    @Autowired
-    lateinit var dbClearer: DBClearer
-
-    companion object {
-        @Container
-        private val postgresqlContainer: PostgreSQLContainer<*> =
-            PostgreSQLContainer<Nothing>("postgres:14.0-alpine").apply {
-                withDatabaseName("foo")
-                withUsername("foo")
-                withPassword("secret")
-            }
-
-        @JvmStatic
-        @DynamicPropertySource
-        fun properties(registry: DynamicPropertyRegistry) {
-            registry.add("spring.datasource.url", postgresqlContainer::getJdbcUrl)
-            registry.add("spring.datasource.password", postgresqlContainer::getPassword)
-            registry.add("spring.datasource.username", postgresqlContainer::getUsername)
-        }
-    }
-
-    @BeforeEach
-    fun clearDB() {
-        dbClearer.clearAll()
-    }
 
     @Test
     @Order(1)
