+++ backend/src/test/kotlin/org/darkSolace/muse/userModule/UserControllerApiTests.kt
@@ -1,29 +1,19 @@
 package org.darkSolace.muse.userModule
 
-import org.darkSolace.muse.DBClearer
 import org.darkSolace.muse.securityModule.model.JwtResponse
 import org.darkSolace.muse.securityModule.model.SignUpRequest
+import org.darkSolace.muse.testUtil.TestBase
 import org.darkSolace.muse.userModule.model.Role
 import org.darkSolace.muse.userModule.model.User
 import org.darkSolace.muse.userModule.service.UserRoleService
 import org.darkSolace.muse.userModule.service.UserService
 import org.junit.jupiter.api.*
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.boot.test.web.client.TestRestTemplate
-import org.springframework.boot.web.server.LocalServerPort
 import org.springframework.http.*
-import org.springframework.test.context.DynamicPropertyRegistry
-import org.springframework.test.context.DynamicPropertySource
-import org.testcontainers.containers.PostgreSQLContainer
-import org.testcontainers.junit.jupiter.Container
-import org.testcontainers.junit.jupiter.Testcontainers
 import java.util.*
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@Testcontainers
-@TestMethodOrder(MethodOrderer.OrderAnnotation::class)
-class UserControllerApiTests {
+class UserControllerApiTests : TestBase() {
     @Autowired
     private lateinit var restTemplate: TestRestTemplate
 
@@ -32,38 +22,6 @@ class UserControllerApiTests {
 
     @Autowired
     private lateinit var userRoleService: UserRoleService
-
-    @LocalServerPort
-    private lateinit var port: String
-
-    @Autowired
-    private lateinit var dbClearer: DBClearer
-
-    companion object {
-        @Container
-        private val postgresqlContainer: PostgreSQLContainer<*> =
-            PostgreSQLContainer<Nothing>("postgres:14.0-alpine").apply {
-                withDatabaseName("foo")
-                withUsername("foo")
-                withPassword("secret")
-            }
-
-        @JvmStatic
-        @DynamicPropertySource
-        fun properties(registry: DynamicPropertyRegistry) {
-            registry.add("spring.datasource.url", postgresqlContainer::getJdbcUrl)
-            registry.add("spring.datasource.password", postgresqlContainer::getPassword)
-            registry.add("spring.datasource.username", postgresqlContainer::getUsername)
-        }
-    }
-
-    @BeforeEach
-    fun clearDB() {
-        dbClearer.clearAll()
-    }
-
-    private fun generateUrl(path: String): String = "http://localhost:$port$path"
-
 
     @Test
     @Order(1)
