+++ backend/src/test/kotlin/org/darkSolace/muse/userModule/UserTagServiceTests.kt
@@ -1,56 +1,21 @@
 package org.darkSolace.muse.userModule
 
-import org.darkSolace.muse.DBClearer
+import org.darkSolace.muse.testUtil.TestBase
 import org.darkSolace.muse.userModule.model.User
 import org.darkSolace.muse.userModule.model.UserTag
 import org.darkSolace.muse.userModule.service.UserService
 import org.darkSolace.muse.userModule.service.UserTagService
 import org.junit.jupiter.api.Assertions
-import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.fail
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.test.context.SpringBootTest
-import org.springframework.test.context.DynamicPropertyRegistry
-import org.springframework.test.context.DynamicPropertySource
-import org.testcontainers.containers.PostgreSQLContainer
-import org.testcontainers.junit.jupiter.Container
-import org.testcontainers.junit.jupiter.Testcontainers
 
-@SpringBootTest
-@Testcontainers
-class UserTagServiceTests {
+class UserTagServiceTests : TestBase() {
     @Autowired
     lateinit var userTagService: UserTagService
 
     @Autowired
     lateinit var userService: UserService
-
-    @Autowired
-    lateinit var dbClearer: DBClearer
-
-    companion object {
-        @Container
-        private val postgresqlContainer: PostgreSQLContainer<*> =
-            PostgreSQLContainer<Nothing>("postgres:14.0-alpine").apply {
-                withDatabaseName("foo")
-                withUsername("foo")
-                withPassword("secret")
-            }
-
-        @JvmStatic
-        @DynamicPropertySource
-        fun properties(registry: DynamicPropertyRegistry) {
-            registry.add("spring.datasource.url", postgresqlContainer::getJdbcUrl)
-            registry.add("spring.datasource.password", postgresqlContainer::getPassword)
-            registry.add("spring.datasource.username", postgresqlContainer::getUsername)
-        }
-    }
-
-    @BeforeEach
-    fun clearDB() {
-        dbClearer.clearAll()
-    }
 
     @Test
     fun addTagToUser() {
@@ -59,8 +24,7 @@ class UserTagServiceTests {
             User(
                 username = "testUser14",
                 password = "123",
-                email = "test14@example.com",
-                userTags = mutableSetOf()
+                email = "test14@example.com"
             )
         ) ?: fail("user is null")
 
@@ -93,8 +57,7 @@ class UserTagServiceTests {
             User(
                 username = "testUser914",
                 password = "123",
-                email = "test914@example.com",
-                userTags = mutableSetOf()
+                email = "test914@example.com"
             )
         )
         val userByUsername = User(username = "testUser914", password = "", email = "")
